refactor VisibilityEvaluator, etc
